Receive websocket updates correctly.
